backend/TODO_list/requirements/dev.txt
coverage==4.4.1
Django==1.11.5
django-cors-headers==2.1.0
djangorestframework==3.6.4
pytz==2017.2
selenium==3.5.0
